Added Double Jumping and Dashing w/ SFX
Added double jumping and dashing (on cooldown) with SFX. Also changed the comments detailing variable clumps to more Unity appropriate "Headers".

It is important that we all read about coroutines and their functionality and uses. I added both a youtube link of the video I got my dashing code and knowledge from as well as a link to Unity's own documentation on coroutines. Both links are in the code in a comment above the dash coroutine.